refactor(object): split 423-line module into object-utils, object-transform, object-access
src/object.ts is now a thin barrel re-exporting from three focused sub-modules
under src/_internal/:

- object-utils.ts     — isObject (shared type guard, no internal deps)
- object-transform.ts — pick, omit, merge, deepMerge, mapValues, mapKeys,
                        filterKeys, filterValues
- object-access.ts    — keys, values, entries, fromEntries, hasKey, getPath,
                        setPath, updatePath, isEmpty, equals, clone, freeze

Moving isObject to its own file also resolves the implicit use-before-definition
smell where deepMerge (line 73) referenced isObject (line 325) in the old file.

All public exports and the fp-core/object subpath import remain unchanged.

Closes #27
